attack(opponent) refactored
attack(opponent) now only exists in DungeonCharacter, this was made possible by implementing Strategy, because now the contents of each attack is stored in the interface Attack, rather than being hard coded in.
While this was indeed made possible with the Strategy refactor, this is a sepperate refactor as its getting rid of duplicate methods in all the subclasses of Hero and Monster.